Correct the set fact to fit the return structure
It seems there was a structure change of what is returned by the module. It's now a list of dict and what was the key become the value of `role_name`. I adapted the example with jinja2 filter to access the privileges list of 'Admin'.

I also adapted the type. PS: I'm not sure about that change. Maybe it should be only `type: list`…