vmware_guest_storage_policy: test: don't clean up the test VM
The test VM will be removed later anyway. By doing this, we avoid the
following error:

"msg": "Unable to find the datastore with given parameters. This could mean, DC0_H0_VM1 is a non-existent virtual machine and module tried to deploy it as new virtual machine with no disk. Please specify disks parameter or specify template to clone from."