Fix resourcepool error and add disk type
Add destination_disk_type to the module parameters. Converts all disks of the virtual machine to 'thick', 'thin', or 'eagerzeroedthick'. This only happens if the disks will move to a different datastore.
Also fixes an issue where the module fails due to destination_resourcepool being required even if destination_host is not specified since this does not follow in line with what the documentation states. Basically we don't need resourcepool at all unless destination_host is set so why do we fail if it isn't found? I removed the failure since it can be interpreted by the host_obj or from the user provided destination_resourcepool. In my opinion we don't need it at all if we are only doing destination_datastore.