tests: ensure we can create 'VM Network'
When we start our tests, VM Network is on vSwitch0. This is
configuration is conflicting when we want to create the same portgroup
on the `isolate_vSwitch0`.
So far, we avoided the conflict because the teardown_with_esxi.yml
playbook was called at the end of the first target and cleaning up the
portgroup.

This is not the case anymore because:
- we randomize¹ the target execution order
- the vmware_cluster_info target does not attach the ESXi host and so
  the teardown was not working as expected.

This commit introduces the following changes:

- alway clean up VM Network from vSwitch0 before we try to recreate
  it on the `isolated_vSwitch0`.
- move vmware_cluster_info in zuul/vmware/vcenter_only where it belongs

¹: well it's a round robin and it remains deterministic