vmware: Change to use from isinstance to type in if condition (#983)
vmware: Change to use from isinstance to type in if condition

Depends-on #991
Depends-on #977
SUMMARY
I think that the isinstance shouldn't use in the evaluation formula for bool because it occurs mistaking evaluation.
For example:
if isinstance(True, int): # <= is judged as a True
    pass
The cause of mistaking evaluation is that bool type is implemented as an integer sub-class.
Please see the following:

https://docs.python.org/3/library/functions.html#isinstance
https://docs.python.org/3/c-api/bool.html

When using the isinstance, the vmware_vcenter_settings module doesn't work well.
Because bool will be converted to an integer by the following processing.

  
    
      community.vmware/plugins/module_utils/vmware.py
    
    
        Lines 1051 to 1052
      in
      634ec3b
    
    
    
    

        
          
           elif isinstance(option_value, int): 
        

        
          
               option_value = VmomiSupport.vmodlTypes['int'](option_value) 
        
    
  


Also see: #981
So, I think that type should use, not the isinstance.
https://docs.python.org/3/library/functions.html#type
This PR will change to use from isinstance to type.
fixes: #980
What do you think?
ISSUE TYPE

Bugfix Pull Request

COMPONENT NAME
changelogs/fragments/983-vmware.yml
plugins/module_utils/vmware.py
ADDITIONAL INFORMATION
tested on vCenter/ESXi 7.0.2

Reviewed-by: Abhijeet Kasurde <None>
Reviewed-by: sky-joker <sky.jokerxx@gmail.com>
Reviewed-by: Mario Lenz <m@riolenz.de>
Reviewed-by: None <None>