vmware_vswitch: Add test for removing a NIC (#1076)
vmware_vswitch: Add test for removing a NIC

SUMMARY
This is a follow-up to #840. This PR added the ability to remove NICs from a vswitch. I think we should have a test case for this to avoid future changes breaking this.
ISSUE TYPE

Feature Pull Request

COMPONENT NAME
vmware_vswitch

Reviewed-by: sky-joker <sky.jokerxx@gmail.com>
Reviewed-by: None <None>