Add option to mark cloned vm as template vmware_guest_cross_vc_clone (#996)
Add option to mark cloned vm as template vmware_guest_cross_vc_clone

With this change it is possible to clone a template and mark it as a template.
SUMMARY

This change adds a new parameter (is_template).
Since this module can also be used to clone templates it makes sense to have an option to also mark it as a template. Otherwise it will create a virtual machine from the cloned vm/template.
I chose is_template as the param name to be consistent with the vmware_guest module.

ISSUE TYPE


Feature Pull Request

COMPONENT NAME

vmware_guest_cross_vc_clone
ADDITIONAL INFORMATION

Reviewed-by: Abhijeet Kasurde <None>
Reviewed-by: Mario Lenz <m@riolenz.de>
Reviewed-by: None <None>