vmware_vm_inventory/tests: only spawn 1 VM with 1GB of RAM (#1130)
vmware_vm_inventory/tests: only spawn 1 VM with 1GB of RAM

Fedora cannot boot properly with just 128MB of memory. This was not a
problem with ESXi 6.7.0U3 but with 7.0.3, the ESXi can figure out that
the VM has crashed because it's run out of for memory.
So, now we start one single VM and we ensure it's got enough memory.

Reviewed-by: None <None>