[backport] vmware_guest: Fix version_added of advanced_settings (#1157)
[backport] vmware_guest: Fix version_added of advanced_settings

Backport of #1156:
SUMMARY
The docs for vmware_guest  incorrectly state that the advanced_settings parameter was added in 1.7.0.
This PR sets the release date to the correct version 1.8.0.
ISSUE TYPE

Docs Pull Request

COMPONENT NAME
vmware_guest
ADDITIONAL INFORMATION
The feature was added in #179 as d5d00e3.
It did not make it into 1.7.0 but was released as part of 1.8.0, see the release notes.

Reviewed-by: None <None>