Fix getting the next release (#1211)
Fix getting the next release

SUMMARY
I think we have a bug creating the new version running tools/prepare_release.yml. At the moment, we use the last tag as the basis for the new version:

  
    
      community.vmware/tools/prepare_release.yml
    
    
        Lines 3 to 7
      in
      89695ca
    
  
  
    

        
          
           - name: Generate version 
        

        
          
             block: 
        

        
          
               - name: Get the next release version 
        

        
          
                 ansible.builtin.shell: git tag --sort=-creatordate|head -n1|perl -pe 's/(\d+\.)(\d+)\.\d+/"$1" . ($2+1) . ".0"/e' 
        

        
          
                 register: result 
        
    
  


Now that the latest tag is 1.17.1, this gives 1.18.0 when used in the main branch instead of 2.1.0.
I think we can fix this by using the --merged parameter. As far as I understand, the 1.17.1 tag wouldn't turn up then because the tagged commit isn't part of the main branch.
ISSUE TYPE

Bugfix Pull Request

COMPONENT NAME
tools/prepare_release.yml
ADDITIONAL INFORMATION
I think the bug would work the other way, too. If we release 2.1.0 and then want to release 1.18.0 in the stable-1.x branch, we would generate 2.2.0.

Reviewed-by: Gonéri Le Bouder <goneri@lebouder.net>