Order uplink ports by uplink name (#1243)
Order uplink ports by uplink name

SUMMARY

When attaching pNics to Distributed Virtual Switches, the vSphere administrator may expect the pNic to be attached to the host uplinks in the named order that is visible in the vSphere Client.   For example, vmnic0 to "Uplink 1", vmnic1 to "Uplink 2" and so on.
The current behavior of the module attaches the first pNic to the first uplink port that it finds for the host, which might not be the first uplink for the host by name.
This change simply sorts the fetched dvSwitch uplink ports for the host by the name of the port.  Then the nics will the be attached in the expected order.
Fixes #1242
Fixes #1070

ISSUE TYPE


Bugfix Pull Request

COMPONENT NAME

vmware_dvs_host
ADDITIONAL INFORMATION

Reviewed-by: Mario Lenz <m@riolenz.de>