fix(vmware_guest): parameter not correct nicsettings:adapter:ip
Usage of Customization Spec, when configured to `prompt user` for Network settings, caused error `Failed to create a virtual machine: A specified parameter was not correct: nicsettings:adapter:ip`.
This fix ensure IP Address, Subnet Mask and Hostname will be set based on parameter `networks` and `name` if parameter `customization_spec` is used.

Refs: https://github.com/ansible-collections/community.vmware/issues/599
Refs: https://github.com/ansible/ansible/issues/55560
Refs: https://github.com/ansible/ansible/pull/61329/