Add support for configuring SR-IOV advanced options on a network interface via vmware_guest_network (#1218)
Add support for configuring SR-IOV advanced options on a network interface via vmware_guest_network

SUMMARY

Port the additions made by pull request #1080 to the vmware_guest_network module so as to add SR-IOV NICs with advanced options independently of VM creation.

ISSUE TYPE

Feature Pull Request

COMPONENT NAME
vmware_guest_network: Add the following options to a network interface definition, which are to be used for a sriov device :

allow_gest_os_mtu_change
physical_function_backing
virtual_function_backing

ADDITIONAL INFORMATION
This is just a port of the code additions to vmware_guest from pull request 1080, to enable pinning a SR-IOV interface to specific hardware when adding/managing interfaces via vmware_guest_network.
I noticed that when creating a VM, and creating all interfaces at the same time, "SR-IOV network adapter" ordering and VM interface ordering would get swapped as follows :

hw_eth0 : Network adapter 1
hw_eth1 : SR-IOV network adapter 2
hw_eth2 : SR-IOV network adapter 1

Given I had a requirement to pin an interface to a specific network card (and the equipment it is connected to), this was a big problem. I noticed this problem did not happen when adding the NICs one by one by hand, therefore, I turned to the vmware_guest_network module in order to add the NICs after creating my VMs, using a loop call as is now recommended.
This is when I noticed that this module did not support the SR-IOV advanced options, and that it was not relying on the helper module for creating NIC specifications, so I ported the appropriate changes.
The following code then makes it possible to get the results I want :
- community.vmware.vmware_guest:
  <<: *vmware_common_parameters
  network_name: "{{ item.name|default(omit) }}"
  device_type: "{{ item.device_type|default(omit) }}"
  physical_function_backing: "{{ item.physical_function_backing|default(omit) }}"
  loop:
    - { 'name': PortGroup, 'device_type': 'sriov', 'physical_function_backing': '0000:5d:00.0' }
    - { 'name': PortGroup, 'device_type': 'sriov', 'physical_function_backing': '0000:5d:00.1' }

Reviewed-by: Mario Lenz <m@riolenz.de>
Reviewed-by: Stéphane Lapie <stephane.lapie@darkbsd.org>