VCSIM removal part 4 (#1294)
VCSIM removal part 4

Depends-On: ansible/ansible-zuul-jobs#1511
SUMMARY
This is a follow-up PR to #1293.
I wanted to remove VCSIM from the Developer Guide but was asked to move it to the collection (ansible/ansible#77585). I'll start with this and then check if VCSIM mentioned somewhere else in the documentation.
ISSUE TYPE

Docs Pull Request

COMPONENT NAME
ADDITIONAL INFORMATION

Reviewed-by: Sandra McCann <samccann@redhat.com>
Reviewed-by: Gonéri Le Bouder <goneri@lebouder.net>
Reviewed-by: Mario Lenz <m@riolenz.de>