vmware_guest_vgpu: manage NVIDIA vGPU profile of virtual machines (#1319)
vmware_guest_vgpu: manage NVIDIA vGPU profile of virtual machines

Replaces the apparently abandoned but much needed #656. In addition to the original PR, this one references a newer module for powering off/on a VM and removes the restriction on accepted vGPU profiles leaving their validation up to the server, as the list will depend on a particular GPU model.
SUMMARY
Modify vGPU video card profile of specified virtual machine in given vCenter infrastructure
This module is used to reconfigure vGPU card profile of a given virtual machine.
On a vSphere environment if you need proper 3D support for basic to complex rendering the best choice is to have an attached GPU to the ESXi host. Nvidia been the lead vGPU supplier for VMware and most documentations are covering how to use and configure such devices. Nvidia has moved from a normal PCI passthrough connection to a vGPU style to enable eh use of its licensing service. A very limited number of GPUs that support PCI passthrough are currently supported. The trend is to use vGPU.
This module will help automate adding and removing a vGPU profile (as in the way the new NVIDIA cards reserve GPU resources to a specific VM) using ansible. I believe this module won't be as popular as the other module in this repo for many reasons (very very expensive GPUs with an expensive licensing schema), but still, some people will needs as I did hence why I use the help of a fellow VMware + NVIDIA GPU user and modified his code to build this one.
ISSUE TYPE

New Module Pull Request

COMPONENT NAME
vmware_guest_vgpu
ADDITIONAL INFORMATION
This module requires the VM to be power off and located within the ESXi host that has the GPU attached.
Currently, it has two functions adding and removing a GPU profile. For my various use cases, I have never used anything other than that.
Sample execution and output:
- hosts: localhost
  gather_facts: no
  tasks:
    - name: Add vGPU profile to VM
      vmware_guest_vgpu:
        hostname: "{{ vcenter_hostname }}"
        username: "{{ vcenter_username }}"
        password: "{{ vcenter_password }}"
        datacenter: "{{ datacenter_name }}"
        validate_certs: no
        name: UbuntuTest
        vgpu: 'grid_m10-8q'
        state: present
      register: vgpu_facts

    - name: Generating vGPU facts logs
      debug: var=vgpu_facts


"pci": {
        "changed": true,
        "failed": false,
        "vgpu": {
            "Controller_Key": 100,
            "Key": 13000,
            "Label": "PCI device 0",
            "Summary": "NVIDIA GRID vGPU grid_m10-8q",
            "Unit_Number": 18,
            "Vgpu": "grid_m10-8q"
        }
    }

Reviewed-by: Mario Lenz <m@riolenz.de>