vmware_host_passthrough: Fix to be able to specify PCI id in the devices parameter (#1365)
vmware_host_passthrough: Fix to be able to specify PCI id in the devices parameter

Depends-On: ansible/ansible-zuul-jobs#1579
SUMMARY
If ESXi has multiple PCI devices and they are the same name, and you want to enable the passthrough, you need to specify the PCI id.
But the module hasn't been able to specify the PCI id of the PCI device.
This PR is to be able to specify the id in the devices parameter the module has.
fixes: #1364
ISSUE TYPE

Feature Pull Request

COMPONENT NAME
plugins/modules/vmware_host_passthrough.py
ADDITIONAL INFORMATION
tested in vCenter/ESXi 7.0.0

Reviewed-by: Mario Lenz <m@riolenz.de>