Improve module documentation (#1416)
Improve module documentation

SUMMARY
Python and pyVmomi requirements are already defined at the collection level. So they are, in the best case, redundant in the modules and, in the worst case (mostly, if not always), conflicting.
Additionally, all those Tested on documentation is quite misleading. A module might have been tested against a specific vCenter or ESXi version originally, but I assume that in most cases changes to the module weren't. Also, the CI tests run with a version not mentioned in the modules.
ISSUE TYPE

Docs Pull Request

COMPONENT NAME
nearly all :-P
ADDITIONAL INFORMATION
#1412

Reviewed-by: Gonéri Le Bouder <goneri@lebouder.net>