vmware_dvs_portgroup : Remove defaults and add explicit inheritance from switch defaults (#1483)
vmware_dvs_portgroup : Remove defaults and add explicit inheritance from switch defaults

SUMMARY
Apart from the things that are specific to a portgroup (like, for example, its name) there are a lot of settings that could either be inherited from the switch or defined at the PG level.
Most, if not all, parameters of the module don't allow this: They have defaults which means the settings from the switch are always overwritten. Or they implicitly set something to inherited if the parameter is unset. This is a problem because people can't use the module in a way of "I don't care about this setting, just leave it as it is" manner or explicitly state that this setting should be inherited.
ISSUE TYPE

Feature Pull Request

COMPONENT NAME
vmware_dvs_portgroup
ADDITIONAL INFORMATION
#1316