Update docs for 3.0.0 (#1493)
Update docs for 3.0.0

I don't know why, but preparing the new release 3.0.0 lead to a lot of changed docs. More than there have been changes to modules, which is a bit weird.
I guess there have been some changes in ansible-core or collection_prep or somewhere, so I'll try to split the changes for 3.0.0 up in two PRs.