This Commit adds support for the script_text Parameter. (#1485)
Added scriptText Parameter to execute Scripts in Linux

SUMMARY

This Commit adds support for the scriptText Parameter.
The scriptText Parameter allows to execute any script with a valid shebang.
It is used in the VM Customization Specifications for the Customization script
https://docs.vmware.com/en/VMware-vSphere/7.0/com.vmware.vsphere.vm_admin.doc/GUID-9A5093A5-C54F-4502-941B-3F9C0F573A39.html

ISSUE TYPE


Feature Pull Request

COMPONENT NAME

vmware_guest

Reviewed-by: Mario Lenz <m@riolenz.de>