Remove timeout parameter in wait_for_task
This was originally added for the Ansible 2.6 release, but in Ansible
2.10 a `timeout` property was added at the task level, which makes this
functionality obsolete.

However, the presence of the `timeout` parameter to this function means
that it effectively overrides the aforementioned task-level property.
This means that all tasks that use `wait_for_task` effectively have a
1-hour timeout, regardless of what the user requests using the `timeout`
property.

Having looked through this collection's source code, it does not seem
that any callers to this function are overriding the `timeout`
parameter, and no unit tests were added either. Thus, the only change
necessary is to this function itself.