fixing ssl validation behaviour (#1653)
fixing ssl validation behaviour

SUMMARY
Fixes #1433
Fixing behaviour of SSL Validation at destination host in cross_vc_clone module. Since it is possible to define to not perform ssl validation at destination VC, there is still an error, because serviceLocatorSpec wants a thumbprint, if this is not set, it throws and ssl validation error.
ISSUE TYPE

Bugfix Pull Request

COMPONENT NAME
vmware_guest_cross_vc_clone
ADDITIONAL INFORMATION
What I am doing to work around the limits of the serviceLocatorSpec, is to check if ssl should be validated, if not, I just fetch the thumbprint from the destination VC and inject it to the serviceLocatorSpec.
There is also another PR currently open (#1608) which adds a thumbprint option, this is also a really good addition. May we should bring my change and the change of @ironbishop and me together, what to you think @mariolenz ?

Reviewed-by: Mario Lenz <m@riolenz.de>