Add switchIpAddress to vmware_dvswitch module (#1949)
Add switchIpAddress to vmware_dvswitch module

SUMMARY
This change makes it possible to change the Switch IP address parameter for the NetFlow config of a distributed vSwitch:

ISSUE TYPE

Feature Pull Request

COMPONENT NAME
vmware_dvswitch
ADDITIONAL INFORMATION
As this is a parameter which belongs to the NetFlow config, I added it under the conditional of the netFlow_collector_ip, if this unwanted I can of course pull it one layer out!
I'm not used to contributing to Modules so far, so please just point out every issue or improvement if needed, I will fix it 👍🏻
Tested with a self-built EE via navigator so far.
Thanks!

Reviewed-by: Mario Lenz <m@riolenz.de>
Reviewed-by: Lukas Magauer