vmware_guest_info - Make "datacenter" property optional (#2241)
SUMMARY

We run a big VMware environment featuring multiple vCenters and multiple datacenters within them.
When automating the deployment of VMs at the moment we search for existing machines with the name of the to be deployed machine in order to perform clean-up activities. At the moment we use the guest find module, but it's slow.
Therefore we looked into the guest_info module and found that if the VM is unique, the datacenter property is never actually used as it is only relevant in case multiple VMs have been found.
In case multiple VMs are found, the folder and datacenter attributes are used to decide on "the right one".
Therefore, the datacenter property could as well be an optional parameter on the module and it could simply fail in case there is more than one VM matching and "datacenter" as well as "folder" are required.

ISSUE TYPE


Feature Pull Request

COMPONENT NAME

vmware_guest_info.py
vmware.py

ADDITIONAL INFORMATION

Reviewed-by: Mario Lenz <m@riolenz.de>