Update vmware_guest.py docs (#2272)
SUMMARY

None of name or vlan is in fact required. For correct operation either one can be specified.

ISSUE TYPE


Docs Pull Request

COMPONENT NAME

vmware_guest
ADDITIONAL INFORMATION


This cost us about 2 hours of debugging because we thought that vlan is required. When we specified both name and vlan, another network was picked than we specified in name because vlan takes precedence and there was another network available with the same vlan id.
This PR might be worth backporting to all version where this applies.
Relevant code section:

  
    
      community.vmware/plugins/modules/vmware_guest.py
    
    
        Lines 1811 to 1812
      in
      e249a03
    
  
  
    

        
          
           self.module.fail_json(msg="Please specify at least a network name or" 
        

        
          
                                     " a VLAN name under VM network list.")

Reviewed-by: Mario Lenz <m@riolenz.de>