vmware_object_role_permission: Fix recursive (#2351)
Fixes #2350
SUMMARY
Currently, the module ignores changes to the recursive option.
ISSUE TYPE

Bugfix Pull Request

COMPONENT NAME
vmware_object_role_permission
ADDITIONAL INFORMATION
This also means that recursive has to be correct when removing a permission. But I think that's OK because it makes sense: If you want to remove a non-recursive permission, it doesn't make sense to remove a recursive  one and vice versa.

Reviewed-by: Alexander Nikitin <ihumster@ihumster.ru>