Fix WebHandle to detect range support without relying on 'accept-ranges' header (#2399)
Some HTTP/2 servers, such as Sonatype Nexus, do not send the accept-ranges header,
even though they support ranged downloads. This causes vmware_deploy_ovf.py to fail
with "Site does not accept ranges", despite the server correctly supporting range requests.
This patch changes the logic to verify range support by sending a Range: bytes=0-0 header
and checking if the response status is 206 Partial Content, which is the actual indicator
of support.
This makes the module compatible with modern HTTP/2 servers such as Nexus and S3.
This change is also fully compatible with HTTP/1.1 servers, which respond identically to Range headers.
Thus, the new logic is both more accurate and backward-compatible.
Fixes: None directly, but observed with Nexus 3.x.

Reviewed-by: Alexander Nikitin <ihumster@ihumster.ru>
Reviewed-by: Mario Lenz <m@riolenz.de>