fix: vmware_host_iscsi now works as expected when there are multiple iscsi devices
SUMMARY

Previously, `vmware_host_iscsi` would pick an arbitrary iSCSI device and
expect it's name to match `vmhba_name`,
which means it would only work as expected 100% of the time if there was only one
iSCSI device.

Now we explicitly search for an iSCSI device named `vmhba_name`, so we won't be
fiddling with the wrong device by mistake.

Since there was already this check later in the code (function `check_hba_name`), this means
this plugin was NOT attempting to modify the wrong device. It was just giving an error
and causing an unecessary failture. I.e. this commit is NOT preventing data loss,
for there wasn't any to begin with.